fixing spacing and responsiveness fro additemview and pagewrapper
