+++ src/components/AddItemForm/additemform.css
@@ -1,0 +1,16 @@
+.add-item label {
+  margin-bottom: 15px;
+}
+
+.input-container {
+  margin-bottom: 35px;
+}
+
+.how-soon-div {
+  text-align: center;
+}
+
+.add-item .orange-button {
+  display: block;
+  margin: 0 auto;
+}
