+++ src/components/Buttons/Buttons.css
@@ -4,7 +4,7 @@
   height: 3.18em;
   background: #ffa001;
   border-radius: 2.438em;
-  font-family: Poppins;
+  font-family: 'Poppins', sans-serif;
   font-size: 1em;
   color: #ffffff;
   border: 0px;
@@ -30,11 +30,12 @@
 .soon-buttons {
   display: flex;
   align-items: center;
-  justify-content: center;
+  justify-content: space-evenly;
   flex-direction: row;
-  font-family: Poppins;
+  font-family: 'Poppins', sans-serif;
   font-weight: 600;
   font-size: 100%;
+  margin: 35px 0;
 }
 
 .soon-buttons label {
