+++ src/components/ItemCard/itemcard.css
@@ -4,7 +4,7 @@
   border-radius: 0px 14px 14px 0px !important;
   border-left: 5px solid;
   padding: 15px;
-  margin-bottom: 25px;
+  margin-bottom: 35px;
 }
 
 .css-1elwnq4-MuiPaper-root-MuiAccordion-root:before {
