+++ src/components/PageTitle/page-title.css
@@ -3,6 +3,7 @@
 }
 
 h1 {
+  margin-top: 0;
   text-align: center;
   position: relative;
 }
