+++ src/components/PageWrapper/pagewrapper.css
@@ -17,7 +17,7 @@
   width: 50%;
   min-height: 80vh;
   margin: 0 auto;
-  padding: 2% min(4%, 40px);
+  padding: 4% min(4%, 40px);
 }
 
 @media only screen and (max-width: 768px) {
@@ -28,7 +28,6 @@
     width: 100%;
     border-radius: 0;
     box-shadow: none;
-    padding-top: 0;
   }
 }
 
@@ -45,5 +44,6 @@
     width: 50%;
     box-shadow: 0px 4px 24px rgba(0, 0, 0, 0.17);
     border-radius: 50px;
+    min-height: 70vh;
   }
 }
