+++ src/components/index.js
@@ -4,7 +4,7 @@ export { default as List } from './List/List';
 export { default as Navigation } from './Navigation/Navigation';
 export { default as PageTitle } from './PageTitle/PageTitle';
 export { default as PageWrapper } from './PageWrapper/PageWrapper';
-export { default as AddItemForm } from './AddItemForm';
+export { default as AddItemForm } from './AddItemForm/AddItemForm';
 export { default as JoinList } from './JoinList';
 export { default as RequireToken } from './RequireToken';
 export { default as RouteLink } from './RouteLink';
